Move all common converters to a separate namespace
