+++ src/Soloplan.WhatsON.CruiseControl.GUI/CruiseControlProjectDataTemplate.xaml
@@ -2,9 +2,11 @@
                     xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
                     xmlns:buildServer="clr-namespace:Soloplan.WhatsON.GUI.Common.BuildServer;assembly=Soloplan.WhatsON.GUI.Common"
                     xmlns:connectorTreeView="clr-namespace:Soloplan.WhatsON.GUI.Common.ConnectorTreeView;assembly=Soloplan.WhatsON.GUI.Common"
-                    xmlns:local="clr-namespace:Soloplan.WhatsON.CruiseControl.GUI;assembly=Soloplan.WhatsON.CruiseControl.GUI">
+                    xmlns:local="clr-namespace:Soloplan.WhatsON.CruiseControl.GUI;assembly=Soloplan.WhatsON.CruiseControl.GUI"
+                    xmlns:common="clr-namespace:Soloplan.WhatsON.GUI.Common;assembly=Soloplan.WhatsON.GUI.Common"
+                    xmlns:converters="clr-namespace:Soloplan.WhatsON.GUI.Common.Converters;assembly=Soloplan.WhatsON.GUI.Common">
 
-  <connectorTreeView:TimeToApproximateTimeConverter x:Key="TimeToApproximateTimeConverter" />
+  <converters:TimeToApproximateTimeConverter x:Key="TimeToApproximateTimeConverter" />
 
   <DataTemplate DataType="{x:Type local:CruiseControlProjectViewModel}">
     <DockPanel Width="Auto" ContextMenu="{DynamicResource BuildServerProjectContextMenu}">
