+++ src/Soloplan.WhatsON.GUI.Common/BuildServer/CulpritsControl.xaml
@@ -5,11 +5,12 @@
              xmlns:local="clr-namespace:Soloplan.WhatsON.GUI.Common.BuildServer"
              xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
              xmlns:common="clr-namespace:Soloplan.WhatsON.GUI.Common"
+             xmlns:converters="clr-namespace:Soloplan.WhatsON.GUI.Common.Converters"
              d:DesignHeight="450"
              d:DesignWidth="800"
              mc:Ignorable="d">
   <UserControl.Resources>
-    <common:CountVisibilityConverter x:Key="CountToVisibility" />
+    <converters:CountVisibilityConverter x:Key="CountToVisibility" />
   </UserControl.Resources>
 
   <StackPanel Visibility="{Binding CulpritsProp, Mode=OneWay, Converter={StaticResource CountToVisibility}}">
