+++ src/Soloplan.WhatsON.GUI.Common/BuildServer/HistoryBuildControl.xaml
@@ -8,13 +8,15 @@
              xmlns:materialDesign="http://materialdesigninxaml.net/winfx/xaml/themes"
              xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
              xmlns:s="clr-namespace:System;assembly=mscorlib"
+             xmlns:common="clr-namespace:Soloplan.WhatsON.GUI.Common"
+             xmlns:converters="clr-namespace:Soloplan.WhatsON.GUI.Common.Converters"
              d:DataContext="{d:DesignInstance local:BuildStatusViewModel}"
              d:DesignHeight="450"
              d:DesignWidth="800"
              ToolTipService.ShowDuration="{x:Static s:Int32.MaxValue}"
              mc:Ignorable="d">
   <UserControl.Resources>
-    <connectorTreeView:StatusToColorConverter x:Key="StatusToColorConverter" />
+    <converters:StatusToColorConverter x:Key="StatusToColorConverter" />
     <ContextMenu x:Key="ContextMenuBuild">
       <MenuItem CommandParameter="{Binding}" Command="{Binding CopyBuildLabel}">
         <MenuItem.Header>
