+++ src/Soloplan.WhatsON.GUI.Common/BuildServer/HistoryBuildTooltipControl.xaml
@@ -6,13 +6,14 @@
              xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
              xmlns:connectorTreeView="clr-namespace:Soloplan.WhatsON.GUI.Common.ConnectorTreeView"
              xmlns:common="clr-namespace:Soloplan.WhatsON.GUI.Common"
+             xmlns:converters="clr-namespace:Soloplan.WhatsON.GUI.Common.Converters"
              d:DataContext="{d:DesignInstance local:BuildStatusViewModel}"
              d:DesignHeight="450"
              d:DesignWidth="800"
              mc:Ignorable="d">
   <UserControl.Resources>
-    <connectorTreeView:TimeToApproximateTimeConverter x:Key="TimeToAproximateTimeConverter" />
-    <common:NullOrDefaultVisibilityConverter x:Key="NullOrDefaultVisibilityConverter" />
+    <converters:TimeToApproximateTimeConverter x:Key="TimeToAproximateTimeConverter" />
+    <converters:NullOrDefaultVisibilityConverter x:Key="NullOrDefaultVisibilityConverter" />
   </UserControl.Resources>
   <StackPanel>
     <StackPanel Orientation="Horizontal" Visibility="{Binding Label, Converter={StaticResource NullOrDefaultVisibilityConverter}}">
