+++ src/Soloplan.WhatsON.GUI.Common/Converters/StatusToColorConverter.cs
@@ -3,12 +3,13 @@
 // Licensed under the MIT License. See License-file in the project root for license information.
 // </copyright>
 
-namespace Soloplan.WhatsON.GUI.Common.ConnectorTreeView
+namespace Soloplan.WhatsON.GUI.Common.Converters
 {
   using System;
   using System.Globalization;
   using System.Windows.Data;
   using System.Windows.Media;
+  using Soloplan.WhatsON.GUI.Common.ConnectorTreeView;
   using Soloplan.WhatsON.Model;
 
   public class StatusToColorConverter : IValueConverter
