+++ src/Soloplan.WhatsON.GUI.Common/Converters/TimeToApproximateTimeConverter.cs
@@ -3,7 +3,7 @@
 // Licensed under the MIT License. See License-file in the project root for license information.
 // </copyright>
 
-namespace Soloplan.WhatsON.GUI.Common.ConnectorTreeView
+namespace Soloplan.WhatsON.GUI.Common.Converters
 {
   using System;
   using System.Globalization;
