+++ src/Soloplan.WhatsON.GUI.Common/Soloplan.WhatsON.GUI.Common.csproj
@@ -49,7 +49,7 @@
       <DependentUpon>BuildNumberControl.xaml</DependentUpon>
     </Compile>
     <Compile Include="BuildServer\CopyBuildLabelCommand.cs" />
-    <Compile Include="CountVisibilityConverter.cs" />
+    <Compile Include="Converters\CountVisibilityConverter.cs" />
     <Compile Include="BuildServer\CulpritsControl.xaml.cs">
       <DependentUpon>CulpritsControl.xaml</DependentUpon>
     </Compile>
@@ -72,15 +72,15 @@
     </Compile>
     <Compile Include="ConnectorTreeView\CustomCommand.cs" />
     <Compile Include="ExternalEnabledStateCommand.cs" />
-    <Compile Include="InvertBoolConverter.cs" />
-    <Compile Include="InvertBoolVisibilityConverter.cs" />
-    <Compile Include="NullOrDefaultVisibilityConverter.cs" />
+    <Compile Include="Converters\InvertBoolConverter.cs" />
+    <Compile Include="Converters\InvertBoolVisibilityConverter.cs" />
+    <Compile Include="Converters\NullOrDefaultVisibilityConverter.cs" />
     <Compile Include="OpenWebPageCommand.cs" />
     <Compile Include="PresentationPlugin.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="ConnectorTreeView\IHandleDoubleClick.cs" />
     <Compile Include="ConnectorTreeView\PluginManagerExtensions.cs" />
-    <Compile Include="ConnectorTreeView\StatusToColorConverter.cs" />
+    <Compile Include="Converters\StatusToColorConverter.cs" />
     <Compile Include="ConnectorTreeView\StatusViewModel.cs" />
     <Compile Include="ConnectorTreeView\ConnectorGroupViewModel.cs" />
     <Compile Include="ConnectorTreeView\ConnectorMissingViewModel.cs" />
@@ -89,7 +89,7 @@
     </Compile>
     <Compile Include="ConnectorTreeView\ConnectorTreeViewModel.cs" />
     <Compile Include="ConnectorTreeView\ConnectorViewModel.cs" />
-    <Compile Include="ConnectorTreeView\TimeToApproximateTimeConverter.cs" />
+    <Compile Include="Converters\TimeToApproximateTimeConverter.cs" />
     <Compile Include="ConnectorTreeView\TreeItemViewModel.cs" />
     <Compile Include="VisualConfig\GroupExpansionSettings.cs" />
     <Compile Include="VisualConfig\ColorSettings.cs" />
