+++ src/Soloplan.WhatsON.Jenkins.GUI/JenkinsProjectDataTemplate.xaml
@@ -2,11 +2,13 @@
                     xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
                     xmlns:buildServer="clr-namespace:Soloplan.WhatsON.GUI.Common.BuildServer;assembly=Soloplan.WhatsON.GUI.Common"
                     xmlns:connectorTreeView="clr-namespace:Soloplan.WhatsON.GUI.Common.ConnectorTreeView;assembly=Soloplan.WhatsON.GUI.Common"
-                    xmlns:gui="clr-namespace:Soloplan.WhatsON.Jenkins.GUI;assembly=Soloplan.WhatsON.Jenkins.GUI">
+                    xmlns:gui="clr-namespace:Soloplan.WhatsON.Jenkins.GUI;assembly=Soloplan.WhatsON.Jenkins.GUI"
+                    xmlns:common="clr-namespace:Soloplan.WhatsON.GUI.Common;assembly=Soloplan.WhatsON.GUI.Common"
+                    xmlns:converters="clr-namespace:Soloplan.WhatsON.GUI.Common.Converters;assembly=Soloplan.WhatsON.GUI.Common">
 
   <BooleanToVisibilityConverter x:Key="BoolToVisibility" />
-  <connectorTreeView:StatusToColorConverter x:Key="StatusToColorConverter" />
-  <connectorTreeView:TimeToApproximateTimeConverter x:Key="TimeToApproximateTimeConverter" />
+  <converters:StatusToColorConverter x:Key="StatusToColorConverter" />
+  <converters:TimeToApproximateTimeConverter x:Key="TimeToApproximateTimeConverter" />
 
   <DataTemplate DataType="{x:Type gui:JenkinsStatusViewModel}">
     <buildServer:HistoryBuildControl DataContext="{Binding}">
